Chore(be) #87: 별점이 0.5 단위로 남겨질 수 있도록 API 변경
- 프론트는 아무리 해도 불가능해서 일단 보류..

+++ be/src/main/java/movlit/be/movie/presentation/dto/response/MovieMyCommentReadResponse.java
@@ -3,6 +3,6 @@ package movlit.be.movie.presentation.dto.response;
 import movlit.be.common.util.ids.MovieCommentId;
 
 public record MovieMyCommentReadResponse(String nickname, String profileImgUrl, MovieCommentId movieCommentId,
-                                         String comment, Integer score) {
+                                         String comment, Double score) {
 
 }
